Removed warning

+++ src/lib/csv/CSVReader.cpp
@@ -50,7 +50,7 @@ std::vector<std::string> CSVReader::readLine(int lineNumber) {
   return row;
 }
 
-std::vector<std::string> CSVReader::searchFor(int columnNumber, std::string content) {
+std::vector<std::string> CSVReader::searchFor(long unsigned int columnNumber, std::string content) {
   std::vector<std::string> row;
 
   if (m_file.is_open()) {
